pyproject.toml
[project]
name = "comfy-ui-daply-nodes"
description = "ComfyUI custom nodes by Daply (image loader from URL, and more)."
version = "1.0.0"
license = { file = "LICENSE.txt" }
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28",
    "Pillow>=9.0",
    "numpy",
    "torch",
]

[project.urls]
Repository = "https://github.com/Daply-AI/comfy_ui_daply_nodes"

[tool.comfy]
PublisherId = "daply"
DisplayName = "Comfy UI Daply Nodes"
Icon = "https://example.com/icon.png"
